# server/common/api.yml
openapi: 3.0.1
info:
  title: Picard Datahub Api
  version: 0.5.2
  description: >-
    The Picard Datahub Api is the public interface between the Datahub and
    Picard
  contact:
    name: Friedrich PICARD GmbH & Co.KG
    url: 'https://www.picard.de'
    email: edv@picard.de
  x-logo:
    url: ''
servers:
  - url: 'https://datahub.picard.de:3500/'
  - url: 'http://localhost:3500/'
paths:
  /api/v1/datahub/products/:
    get:
      tags:
        - Datahub
      parameters:
        - name: offset
          description: >-
            The number of items to skip before starting to collect the result
            set
          schema:
            type: integer
          in: query
        - name: limit
          description: The numbers of items to return
          schema:
            type: integer
          in: query
        - name: datefilter
          description: "Filters only records with changes after specified date (RFC3339), example: 2020-07-21T17:32:28Z"
          schema:
            type: string
            format: date-time
          in: query
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
          description: Responses with a list of products
          x-oad-type: response
      operationId: Receive Product List
      summary: Get all products
      description: This Routes transmit all products Picard is offering
    x-oad-type: operation
  /api/v1/datahub/products/lineartech:
    get:
      tags:
        - Datahub
      parameters:
        - name: offset
          description: >-
            The number of items to skip before starting to collect the result
            set
          schema:
            type: integer
          in: query
        - name: limit
          description: The numbers of items to return
          schema:
            type: integer
          in: query
        - name: datefilter
          description: "Filters only records with changes after specified date (RFC3339), example: 2020-07-21T17:32:28Z"
          schema:
            type: string
            format: date-time
          in: query
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LineartechProduct'
          description: Responses with a list of lineartech products
          x-oad-type: response
      operationId: Receive LinearTech Product List
      summary: Get all  lineartech products
      description: This Routes transmit all lineartech products Picard is offering
    x-oad-type: operation
  /api/v1/datahub/products/quantities/lineartech:
    get:
      tags:
        - Datahub
      parameters:
        - name: offset
          description: >-
            The number of items to skip before starting to collect the result
            set
          schema:
            type: integer
          in: query
        - name: limit
          description: The numbers of items to return
          schema:
            type: integer
          in: query
        - name: datefilter
          description: "Filters only records with changes after specified date (RFC3339), example: 2020-07-21T17:32:28Z"
          schema:
            type: string
            format: date-time
          in: query
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LineartechProductQuantity'
          description: Responses with lineartech stockamount
          x-oad-type: response
      operationId: Receive LinearTech Product Quantities
      summary: Get only  lineartech quantities (smaller)
      description: This Routes transmit all lineartech product quantities Picard is offering
    x-oad-type: operation
  /api/v1/datahub/products/quantities/:
    get:
      tags:
        - Datahub
      parameters:
        - name: offset
          description: >-
            The number of items to skip before starting to collect the result
            set
          schema:
            type: integer
          in: query
        - name: limit
          description: The numbers of items to return
          schema:
            type: integer
          in: query
        - name: datefilter
          description: "Filters only records with changes after specified date (RFC3339), example: 2020-07-21T17:32:28Z"
          schema:
            type: string
            format: date-time
          in: query
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductQuantity'
          description: Responses with a list of products
          x-oad-type: response
      operationId: Receive Product Quantity List
      summary: Get all product quantities
      description: This Routes transmit all product quantities
    x-oad-type: operation
...
  /api/v1/datahub/products/prices/ekugellager:
    get:
      tags:
        - Datahub
      parameters:
        - name: offset
          description: >-
            The number of items to skip before starting to collect the result
            set
          schema:
            type: integer
          in: query
        - name: limit
          description: The numbers of items to return
          schema:
            type: integer
          in: query
        - name: datefilter
          description: "Filters only records with changes after specified date (RFC3339), example: 2020-07-21T17:32:28Z"
          schema:
            type: string
            format: date-time
          in: query
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EkuPrice'
          description: Responses with a list of products
          x-oad-type: response
      operationId: Receive Product Price List Eku
      summary: Get all product prices for ekugellager
      description: This Routes transmit all product prices including all price profiles
    x-oad-type: operation
  /api/v1/datahub/products/prices/pdo:
    get:
      tags:
        - Datahub
      parameters:
        - name: offset
          description: >-
            The number of items to skip before starting to collect the result
            set
          schema:
            type: integer
          in: query
        - name: limit
          description: The numbers of items to return
          schema:
            type: integer
          in: query
        - name: datefilter
          description: "Filters only records with changes after specified date (RFC3339), example: 2020-07-21T17:32:28Z"
          schema:
            type: string
            format: date-time
          in: query
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductPrice'
          description: Responses with a list of products
          x-oad-type: response
      operationId: Receive Product Price Profiles
      summary: Get all product price profiles
      description: This Routes transmit all product prices including all price profiles
    x-oad-type: operation
  '/api/v1/datahub/products/{productid}/quantity/':
    get:
      tags:
        - Datahub
      parameters:
        - name: productid
          schema:
            type: string
            x-oad-type: parameter
          in: path
          required: true
          x-oad-type: parameter
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductQuantity'
          description: Responses with a list of products
          x-oad-type: response
      operationId: Receive Quantity for ProductId
      summary: Get a specific Quantity
      description: This Routes transmit stock quantity
    x-oad-type: operation
  /api/v1/datahub/products/categories/:
    get:
      tags:
        - Datahub
      parameters:
        - name: offset
          description: >-
            The number of items to skip before starting to collect the result
            set
          schema:
            type: integer
          in: query
        - name: limit
          description: The numbers of items to return
          schema:
            type: integer
          in: query
        - name: datefilter
          description: "Filters only records with changes after specified date (RFC3339), example: 2020-07-21T17:32:28Z"
          schema:
            type: string
            format: date-time
          in: query
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductCategory'
          description: Responses with a list of product categories
          x-oad-type: response
      operationId: Receive Product Category List
      summary: Get all product categories
      description: This Routes sends you all products Picard is offering
    x-oad-type: operation
...
components:
  schemas:
    Product:
      type: object
      properties:
        ProductId:
          format: integer
          type: number
        ProductName:
          type: string
        ProductName2:
          type: string
        CategoryId:
          format: integer
          type: number
        CategoryIdEku:
          format: integer
          type: number
        Ean:
          type: array
          items:
            type: string
        ManufacturerId:
          format: integer
          type: number
        ExternManufacturerId:
          format: integer
          type: number
        Attributes:
          type: array
          items:
            type: object
            properties:
              LocalizationIso:
                title: Language Code Iso String
                type: string
              AttributeKey:
                type: string
              AttributeValue:
                type: string
        BuildProducts:
          type: array
          items:
            title: Identical related Products
            type: string
        RelatedProducts:
          type: array
          items:
            title: Identical related Products
            type: string
        SetProducts:
          type: array
          items:
            title: Identical related Products
            type: string
        Dimension:
          type: object
          properties:
            Size:
              type: object
              properties:
                InnerDiameter:
                  type: number
                OuterDiameter:
                  type: number
                Width:
                  type: number
            Weight:
              type: number
        Images:
          type: object
          properties:
            Ekugellager:
              type: object
              properties:
                FrontImage:
                  type: string
                BackImage:
                  type: string
                AllroundImage:
                  type: string
            Picard:
              type: object
              properties:
                FrontImage:
                  type: string
                BackImage:
                  type: string
                AllroundImage:
                  type: string
                SketchImage:
                  type: string
        SearchTags1:
          type: string
        SearchTags2:
          type: string
        SearchTags3:
          type: string
        CreatedAt:
          type: string
          format: date-time
        ModifiedAt:
          type: string
          format: date-time
        DeletedAt:
          type: string
          format: date-time
    ProductCategory:
      type: object
      properties:
        CategoryId:
          type: number
        CategoryParentId:
          type: number
        Translations:
          type: array
          items:
            type: object
            properties:
              LocalizationIso:
                type: string
              CategoryName:
                type: string
        CreatedAt:
          type: string
          format: date-time
        ModifiedAt:
          type: string
          format: date-time
        DeletedAt:
          type: string
          format: date-time
    ProductQuantity:
      type: object
      properties:
        ProductId:
          type: number
        Quantity:
          type: number
        CreatedAt:
          type: string
          format: date-time
        ModifiedAt:
          type: string
          format: date-time
        DeletedAt:
          type: string
          format: date-time
...
    ProductPrice:
      type: object
      properties:
        ProductId:
          title: Zugehörige Produkt Id
          type: number
        PriceProfile:
          title: Zugehöriges Preis Profil
          type: number
        Currency:
          title: Währung ist immer EUR
          type: string
        ProductPrice:
          title: Produkt Preis in Euro
          type: number
        ListPrice:
          title: offizieller Listenpreis
          type: number
        FictiveListPrice:
          title: falls nicht offizieller Listenpreis
          type: number
        OccasionPrice:
          title: Kunden Occasionspreis
          type: number
        SelectionPrice:
          title: Kunden Selektionspreis
          type: number
        CreatedAt:
          title: Date of creation as ISO String
          type: string
          format: date-time
        ModifiedAt:
          title: Date of modification as ISO String
          type: string
          format: date-time
        DeletedAt:
          title: Date of delete as ISO String
          type: string
          format: date-time
...
    EkuPrice:
      type: object
      properties:
        ProductId:
          type: number
        Range:
          type: object
          properties:
            FromQty:
              title: Preise von
              type: number
            ToQty:
              title: Preise bis
              type: number
            Price:
              type: string
        CreatedAt:
          title: Date of creation as ISO String
          type: string
          format: date-time
        ModifiedAt:
          title: Date of modification as ISO String
          type: string
          format: date-time
        DeletedAt:
          title: Date of delete as ISO String
          type: string
          format: date-time
...
    LineartechProduct:
      type: object
      properties:
        ProductId:
          type: number
        Type:
          type: number
        Length:
          type: number
        MinLengthLeft:
          type: number
        MaxLengthLeft:
          type: number
        LengthToHole:
          type: number
        Height:
          type: number
        Width:
          type: number
        HoleDiameter:
          type: number
        Thread:
          type: number
        NumberHoles:
          type: number
        Sketch:
          type: string
        SchemaFilter:
          type: string
        Weight:
          type: number
        Details:
          type: object
          properties:
            LengthOriginal:
              type: number
            LengthInStock:
              type: number
            LengthLeft:
              type: number
            LengthRight:
              type: number
            CreatedAt:
              type: string
              format: date-time
            ModifiedAt:
              type: string
              format: date-time
            DeletedAt:
              type: string
              format: date-time
        CreatedAt:
          type: string
          format: date-time
        ModifiedAt:
          type: string
          format: date-time
        DeletedAt:
          type: string
          format: date-time
    LineartechProductQuantity:
      type: object
      properties:
        ProductId:
          type: number
        Details:
          type: object
          properties:
            LengthOriginal:
              type: number
            LengthInStock:
              type: number
            LengthLeft:
              type: number
            LengthRight:
              type: number
            CreatedAt:
              type: string
              format: date-time
            ModifiedAt:
              type: string
              format: date-time
            DeletedAt:
              type: string
              format: date-time
  securitySchemes:
    Development:
      type: apiKey
      description: This Api Key is for accessing developer routes
      name: token
      in: header
...
security:
  - Development: []
...